requirements.txt
crewai
crewai_tools
python-decouple
streamlit